Add auto superuser creation
This add a superuser creation migration when bringing up the server, mainlly for dev purposes.
The flow is when migrations are loaded a function in createsuperuser migration file runs and trying to create a superuser.This operation will successeed only once when the superuser is not been created. on every else time the function will be failed and continue.

Signed-off-by: Tal Koren <talko4245@gmail.com>