Add robust filter method to event model
This add a custom filter method that can excute complexed queries.
The method allows queries that can filter by category, location, event size and event starting time.
Category and location can get multipule values(by list) and the result of a multipule values parameter will be the union of rows that match to one of the values.
if more than one parameter was passed to the method, the queryset that will be returned is the intersection between the different parameters lookups.
Event size and start time can get only single values. Those prameters are ment to sreach events which thier matching field are in the range specified. those parameters are passed by a form of a tuple. The first value is the value to look by and the second is  which side of the value :
True = greather than , False = less or equal to

Paramters combination in short :
	multipule values for one parameter = union
	multipule parameters = intersection

Implmentation
* That was implemented by the model manager and queryset mechanism of Django
* Several use case test were added

Signed-off-by: Tal Koren <talko4245@gmail.com>

Split location filter parameter in event search method

This split the search by location into 2 parameters: location name and location city
which are the main parameters to search by. like before the split, both can get more then one value and the fuction behave the same as with other multi value parameters.

Tests:
change location1 fixture and add test filter by city and location name

Signed-off-by: Tal Koren <talko4245@gmail.com>

Remove unnecessary db arguments in event tests

Signed-off-by: Tal Koren <talko4245@gmail.com>