Add create event feature
This add the create event feature.
Create a view with url path, assuming that a user is already logged in and his/her id is in the url path.
Also create a form to gether the data correctly and save the event in DB.

Note:
After successful creation there is a rediraction to present the created event. A template was added only to test the correction of the rediraction and that the event is been saved(it should be extanded later)

Signed-off-by: Tal Koren <talko4245@gmail.com>